+++ src/hkimjp/konpy2/util.clj
@@ -54,3 +54,9 @@
 (defn iso
   [tm]
   (jt/format "YYYY-MM-dd HH:mm:ss" tm))
+
+(defn HH:mm
+  [tm]
+  (jt/format "HH:mm" tm))
+
+; (HH:mm (jt/local-date-time))